Use Binding for DiffBinaryKind since it's kinda unsafe
